Rename CdpSwaggerTabularService and remove PowerFxConfig in InitAsync
