+++ src/tests/Microsoft.PowerFx.Connectors.Tests/PublicSurfaceTests.cs
@@ -51,7 +51,7 @@ namespace Microsoft.PowerFx.Connector.Tests
               "Microsoft.PowerFx.Connectors.RuntimeConfigExtensions",
               "Microsoft.PowerFx.Connectors.RuntimeConnectorContextExtensions",
               "Microsoft.PowerFx.Connectors.SupportsConnectorErrors",
-              "Microsoft.PowerFx.Connectors.Tabular.CdpSwaggerTabularService",
+              "Microsoft.PowerFx.Connectors.Tabular.ConnectorTable",
               "Microsoft.PowerFx.Connectors.Tabular.CdpTabularService",
               "Microsoft.PowerFx.Connectors.Tabular.TabularService",
               "Microsoft.PowerFx.Connectors.Visibility"
